+++ exporters/ostream/include/opentelemetry/exporters/ostream/log_exporter.h
@@ -61,6 +61,9 @@ private:
   bool is_shutdown_ = false;
   mutable opentelemetry::common::SpinLockMutex lock_;
   bool isShutdown() const noexcept;
+  void printAttributes(
+      const std::unordered_map<std::string, opentelemetry::sdk::common::OwnedAttributeValue> &map,
+      const std::string prefix = "\n\t");
 };
 }  // namespace logs
 }  // namespace exporter