+++ exporters/ostream/include/opentelemetry/exporters/ostream/span_exporter.h
@@ -58,68 +58,6 @@ private:
   // Mapping status number to the string from api/include/opentelemetry/trace/canonical_code.h
   std::map<int, std::string> statusMap{{0, "Unset"}, {1, "Ok"}, {2, "Error"}};
 
-  /*
-    print_value is used to print out the value of an attribute within a vector.
-    These values are held in a variant which makes the process of printing them much more
-    complicated.
-  */
-
-  template <typename T>
-  void print_value(const T &item)
-  {
-    sout_ << item;
-  }
-
-  template <typename T>
-  void print_value(const std::vector<T> &vec)
-  {
-    sout_ << '[';
-    size_t i  = 1;
-    size_t sz = vec.size();
-    for (auto v : vec)
-    {
-      sout_ << v;
-      if (i != sz)
-        sout_ << ',';
-      i++;
-    };
-    sout_ << ']';
-  }
-
-// Prior to C++14, generic lambda is not available so fallback to functor.
-#if __cplusplus < 201402L
-
-  class OwnedAttributeValueVisitor
-  {
-  public:
-    OwnedAttributeValueVisitor(OStreamSpanExporter &exporter) : exporter_(exporter) {}
-
-    template <typename T>
-    void operator()(T &&arg)
-    {
-      exporter_.print_value(arg);
-    }
-
-  private:
-    OStreamSpanExporter &exporter_;
-  };
-
-#endif
-
-  void print_value(const opentelemetry::sdk::common::OwnedAttributeValue &value)
-  {
-#if __cplusplus < 201402L
-    opentelemetry::nostd::visit(OwnedAttributeValueVisitor(*this), value);
-#else
-    opentelemetry::nostd::visit(
-        [this](auto &&arg) {
-          /* explicit this is needed by some gcc versions (observed with v5.4.0)*/
-          this->print_value(arg);
-        },
-        value);
-#endif
-  }
-
   // various print helpers
   void printAttributes(
       const std::unordered_map<std::string, opentelemetry::sdk::common::OwnedAttributeValue> &map,