+++ sdk/include/opentelemetry/sdk/metrics/state/async_metric_storage.h
@@ -27,12 +27,15 @@ class AsyncMetricStorage : public MetricStorage
 public:
   AsyncMetricStorage(InstrumentDescriptor instrument_descriptor,
                      const AggregationType aggregation_type,
-                     void (*measurement_callback)(opentelemetry::metrics::ObserverResult<T> &),
-                     const AttributesProcessor *attributes_processor)
+                     void (*measurement_callback)(opentelemetry::metrics::ObserverResult<T> &,
+                                                  void *),
+                     const AttributesProcessor *attributes_processor,
+                     void *state = nullptr)
       : instrument_descriptor_(instrument_descriptor),
         aggregation_type_{aggregation_type},
         measurement_collection_callback_{measurement_callback},
         attributes_processor_{attributes_processor},
+        state_{state},
         cumulative_hash_map_(new AttributesHashMap()),
         temporal_metric_storage_(instrument_descriptor)
   {}
@@ -46,7 +49,7 @@ public:
     opentelemetry::sdk::metrics::ObserverResult<T> ob_res(attributes_processor_);
 
     // read the measurement using configured callback
-    measurement_collection_callback_(ob_res);
+    measurement_collection_callback_(ob_res, state_);
     std::shared_ptr<AttributesHashMap> delta_hash_map(new AttributesHashMap());
     // process the read measurements - aggregate and store in hashmap
     for (auto &measurement : ob_res.GetMeasurements())
@@ -79,8 +82,9 @@ public:
 private:
   InstrumentDescriptor instrument_descriptor_;
   AggregationType aggregation_type_;
-  void (*measurement_collection_callback_)(opentelemetry::metrics::ObserverResult<T> &);
+  void (*measurement_collection_callback_)(opentelemetry::metrics::ObserverResult<T> &, void *);
   const AttributesProcessor *attributes_processor_;
+  void *state_;
   std::unique_ptr<AttributesHashMap> cumulative_hash_map_;
   TemporalMetricStorage temporal_metric_storage_;
 };