+++ sdk/test/metrics/sync_instruments_test.cc
@@ -9,6 +9,8 @@
 #  include "opentelemetry/sdk/metrics/state/multi_metric_storage.h"
 
 #  include <gtest/gtest.h>
+#  include <cmath>
+#  include <limits>
 
 using namespace opentelemetry;
 using namespace opentelemetry::sdk::instrumentationlibrary;
@@ -103,7 +105,7 @@ TEST(SyncInstruments, LongHistogram)
   std::unique_ptr<WritableMetricStorage> metric_storage(new MultiMetricStorage());
   LongHistogram counter(instrument_descriptor, std::move(metric_storage));
   EXPECT_NO_THROW(counter.Record(10l, opentelemetry::context::Context{}));
-  EXPECT_NO_THROW(counter.Record(10l, opentelemetry::context::Context{}));
+  EXPECT_NO_THROW(counter.Record(-10l, opentelemetry::context::Context{}));  // This is ignored
 
   EXPECT_NO_THROW(counter.Record(
       10l, opentelemetry::common::KeyValueIterableView<M>({{"abc", "123"}, {"xyz", "456"}}),
@@ -120,7 +122,11 @@ TEST(SyncInstruments, DoubleHistogram)
   std::unique_ptr<WritableMetricStorage> metric_storage(new MultiMetricStorage());
   DoubleHistogram counter(instrument_descriptor, std::move(metric_storage));
   EXPECT_NO_THROW(counter.Record(10.10, opentelemetry::context::Context{}));
-  EXPECT_NO_THROW(counter.Record(10.10, opentelemetry::context::Context{}));
+  EXPECT_NO_THROW(counter.Record(-10.10, opentelemetry::context::Context{}));  // This is ignored.
+  EXPECT_NO_THROW(counter.Record(std::numeric_limits<double>::quiet_NaN(),
+                                 opentelemetry::context::Context{}));  // This is ignored too
+  EXPECT_NO_THROW(counter.Record(std::numeric_limits<double>::infinity(),
+                                 opentelemetry::context::Context{}));  // This is ignored too
 
   EXPECT_NO_THROW(counter.Record(
       10.10, opentelemetry::common::KeyValueIterableView<M>({{"abc", "123"}, {"xyz", "456"}}),